+++ rs/nervous_system/integration_tests/src/pocket_ic_helpers.rs
@@ -65,7 +65,7 @@ use itertools::Itertools;
 use maplit::btreemap;
 use pocket_ic::{
     management_canister::CanisterSettings, nonblocking::PocketIc, ErrorCode, PocketIcBuilder,
-    UserError, WasmResult,
+    RejectResponse,
 };
 use prost::Message;
 use rust_decimal::prelude::ToPrimitive;
@@ -838,15 +838,6 @@ pub mod nns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(reply) => reply,
-                WasmResult::Reject(reject) => {
-                    panic!(
-                        "list_neurons was rejected by the NNS governance canister: {:#?}",
-                        reject
-                    )
-                }
-            };
             Decode!(&result, ListNeuronsResponse).unwrap()
         }
 
@@ -871,10 +862,6 @@ pub mod nns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to (NNS) manage_neuron failed: {:#?}", s),
-            };
             Decode!(&result, ManageNeuronResponse).unwrap()
         }
 
@@ -913,7 +900,7 @@ pub mod nns {
             pocket_ic: &PocketIc,
             proposal_id: u64,
             sender: PrincipalId,
-        ) -> Result<ProposalInfo, UserError> {
+        ) -> Result<ProposalInfo, RejectResponse> {
             pocket_ic
                 .query_call(
                     GOVERNANCE_CANISTER_ID.into(),
@@ -922,17 +909,7 @@ pub mod nns {
                     Encode!(&proposal_id).unwrap(),
                 )
                 .await
-                .map(|result| match result {
-                    WasmResult::Reply(reply) => {
-                        Decode!(&reply, Option<ProposalInfo>).unwrap().unwrap()
-                    }
-                    WasmResult::Reject(reject) => {
-                        panic!(
-                            "get_proposal_info was rejected by the NNS governance canister: {:#?}",
-                            reject
-                        )
-                    }
-                })
+                .map(|result| Decode!(&result, Option<ProposalInfo>).unwrap().unwrap())
         }
 
         pub async fn wait_for_proposal_execution(
@@ -956,7 +933,7 @@ pub mod nns {
                         // more attempts to get the proposal info to find out if the proposal
                         // actually got executed.
                         let is_benign = [ErrorCode::CanisterStopped, ErrorCode::CanisterStopping]
-                            .contains(&user_error.code);
+                            .contains(&user_error.error_code);
                         if is_benign {
                             continue;
                         } else {
@@ -1004,12 +981,6 @@ pub mod nns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => {
-                    panic!("Call to get_neurons_fund_audit_info failed: {:#?}", s)
-                }
-            };
             Decode!(&result, GetNeuronsFundAuditInfoResponse).unwrap()
         }
 
@@ -1056,16 +1027,6 @@ pub mod nns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(reply) => reply,
-                WasmResult::Reject(reject) => {
-                    panic!(
-                        "get_network_economics_parameters was rejected by the NNS governance \
-                        canister: {:#?}",
-                        reject
-                    )
-                }
-            };
             Decode!(&result, NetworkEconomics).unwrap()
         }
     }
@@ -1097,10 +1058,6 @@ pub mod nns {
         ) -> Result<Nat, TransferError> {
             let call_id = icrc1_transfer_request(pocket_ic, sender, transfer_arg).await;
             let result = pocket_ic.await_call(call_id).await.unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to icrc1_transfer failed: {:#?}", s),
-            };
             Decode!(&result, Result<Nat, TransferError>).unwrap()
         }
 
@@ -1117,10 +1074,6 @@ pub mod nns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to account_balance failed: {:#?}", s),
-            };
             Decode!(&result, Tokens).unwrap()
         }
 
@@ -1156,7 +1109,7 @@ pub mod nns {
 
             // Assert result is ok.
             match result {
-                Ok(WasmResult::Reply(_reply)) => (), // Ok,
+                Ok(_reply) => (), // Ok,
                 _ => panic!("{:?}", result),
             }
         }
@@ -1185,12 +1138,6 @@ pub mod nns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => {
-                    panic!("Call to get_deployed_sns_by_proposal_id failed: {:#?}", s)
-                }
-            };
             Decode!(&result, GetDeployedSnsByProposalIdResponse).unwrap()
         }
 
@@ -1205,10 +1152,6 @@ pub mod nns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to get_wasm failed: {:#?}", s),
-            };
             Decode!(&result, GetWasmResponse)
                 .unwrap()
                 .wasm
@@ -1231,12 +1174,6 @@ pub mod nns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => {
-                    panic!("Call to get_latest_sns_version failed: {:#?}", s)
-                }
-            };
             let response = Decode!(&result, ListUpgradeStepsResponse).unwrap();
             let latest_version = response
                 .steps
@@ -1461,10 +1398,6 @@ pub mod sns {
                 )
                 .await
                 .expect("Error calling manage_neuron");
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to (SNS) manage_neuron failed: {:#?}", s),
-            };
             Decode!(&result, sns_pb::ManageNeuronResponse).unwrap()
         }
 
@@ -1538,7 +1471,7 @@ pub mod sns {
                     Ok(proposal) => proposal,
                     Err(user_error) => {
                         if [ErrorCode::CanisterStopped, ErrorCode::CanisterStopping]
-                            .contains(&user_error.code)
+                            .contains(&user_error.error_code)
                         {
                             continue;
                         } else {
@@ -1573,7 +1506,7 @@ pub mod sns {
             canister_id: PrincipalId,
             proposal_id: sns_pb::ProposalId,
             sender: PrincipalId,
-        ) -> Result<sns_pb::GetProposalResponse, UserError> {
+        ) -> Result<sns_pb::GetProposalResponse, RejectResponse> {
             pocket_ic
                 .query_call(
                     canister_id.into(),
@@ -1585,17 +1518,7 @@ pub mod sns {
                     .unwrap(),
                 )
                 .await
-                .map(|result| match result {
-                    WasmResult::Reply(reply) => {
-                        Decode!(&reply, sns_pb::GetProposalResponse).unwrap()
-                    }
-                    WasmResult::Reject(reject) => {
-                        panic!(
-                            "get_proposal was rejected by the SNS governance canister: {:#?}",
-                            reject
-                        )
-                    }
-                })
+                .map(|result| Decode!(&result, sns_pb::GetProposalResponse).unwrap())
         }
 
         pub async fn list_neurons(
@@ -1611,15 +1534,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(reply) => reply,
-                WasmResult::Reject(reject) => {
-                    panic!(
-                        "list_neurons was rejected by the SNS governance canister: {:#?}",
-                        reject
-                    )
-                }
-            };
             Decode!(&result, sns_pb::ListNeuronsResponse).unwrap()
         }
 
@@ -1784,10 +1698,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to get_neuron failed: {:#?}", s),
-            };
             let response = Decode!(&result, sns_pb::GetNeuronResponse).unwrap();
             match response.result.expect("No result in response") {
                 get_neuron_response::Result::Error(e) => Err(e),
@@ -1915,10 +1825,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to ledger_id failed: {:#?}", s),
-            };
             Decode!(&result, PrincipalId).unwrap()
         }
 
@@ -1932,10 +1838,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to status failed: {:#?}", s),
-            };
             Decode!(&result, Status).unwrap()
         }
 
@@ -1961,10 +1863,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to get_blocks failed: {:#?}", s),
-            };
             Decode!(&result, GetBlocksResponse).unwrap()
         }
 
@@ -2021,10 +1919,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to icrc1_total_supply failed: {:#?}", s),
-            };
             Decode!(&result, Nat).unwrap()
         }
 
@@ -2042,10 +1936,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to icrc1_balance_of failed: {:#?}", s),
-            };
             Decode!(&result, Nat).unwrap()
         }
 
@@ -2075,10 +1965,6 @@ pub mod sns {
             let call_id =
                 icrc1_transfer_request(pocket_ic, canister_id, sender, transfer_arg).await;
             let result = pocket_ic.await_call(call_id).await.unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to icrc1_transfer failed: {:#?}", s),
-            };
             Decode!(&result, Result<Nat, TransferError>).unwrap()
         }
 
@@ -2104,10 +1990,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to get_blocks failed: {:#?}", s),
-            };
             Decode!(&result, GetBlocksResponse).unwrap()
         }
 
@@ -2194,10 +2076,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to archives failed: {:#?}", s),
-            };
             Decode!(&result, Vec<ArchiveInfo>).unwrap()
         }
 
@@ -2216,10 +2094,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to icrc2_approve failed: {:#?}", s),
-            };
             Decode!(&result, Result<Nat, ApproveError>).unwrap()
         }
 
@@ -2238,10 +2112,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to icrc2_allowance failed: {:#?}", s),
-            };
             Decode!(&result, Allowance).unwrap()
         }
 
@@ -2260,10 +2130,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to icrc2_transfer_from failed: {:#?}", s),
-            };
             Decode!(&result, Result<Nat, TransferFromError>).unwrap()
         }
     }
@@ -2295,10 +2161,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to get_blocks failed: {:#?}", s),
-            };
             Decode!(&result, BlockRange).unwrap()
         }
     }
@@ -2382,12 +2244,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => {
-                    panic!("Call to get_sns_canisters_summary failed: {:#?}", s)
-                }
-            };
             Decode!(&result, GetSnsCanistersSummaryResponse).unwrap()
         }
     }
@@ -2512,10 +2368,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to new_sale_ticket failed: {:#?}", s),
-            };
             Decode!(&result, GetInitResponse).unwrap()
         }
 
@@ -2537,10 +2389,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to new_sale_ticket failed: {:#?}", s),
-            };
             Decode!(&result, ListSnsNeuronRecipesResponse).unwrap()
         }
 
@@ -2563,10 +2411,6 @@ pub mod sns {
                 )
                 .await
                 .map_err(|err| err.to_string())?;
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to new_sale_ticket failed: {:#?}", s),
-            };
             Ok(Decode!(&result, NewSaleTicketResponse).unwrap())
         }
 
@@ -2589,10 +2433,6 @@ pub mod sns {
                 )
                 .await
                 .map_err(|err| err.to_string())?;
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to refresh_buyer_tokens failed: {:#?}", s),
-            };
             Ok(Decode!(&result, RefreshBuyerTokensResponse).unwrap())
         }
 
@@ -2613,10 +2453,6 @@ pub mod sns {
                 )
                 .await
                 .map_err(|err| err.to_string())?;
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to get_buyer_state failed: {:#?}", s),
-            };
             Ok(Decode!(&result, GetBuyerStateResponse).unwrap())
         }
 
@@ -2634,10 +2470,6 @@ pub mod sns {
                 )
                 .await
                 .map_err(|err| err.to_string())?;
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to get_open_ticket failed: {:#?}", s),
-            };
             Ok(Decode!(&result, GetOpenTicketResponse).unwrap())
         }
 
@@ -2658,10 +2490,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to error_refund_icp failed: {:#?}", s),
-            };
             Decode!(&result, ErrorRefundIcpResponse).unwrap()
         }
 
@@ -2678,10 +2506,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to get_derived_state failed: {:#?}", s),
-            };
             Decode!(&result, GetDerivedStateResponse).unwrap()
         }
 
@@ -2698,10 +2522,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to get_lifecycle failed: {:#?}", s),
-            };
             Decode!(&result, GetLifecycleResponse).unwrap()
         }
 
@@ -2848,10 +2668,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => panic!("Call to finalize_swap failed: {:#?}", s),
-            };
             Decode!(&result, FinalizeSwapResponse).unwrap()
         }
 
@@ -2868,12 +2684,6 @@ pub mod sns {
                 )
                 .await
                 .unwrap();
-            let result = match result {
-                WasmResult::Reply(result) => result,
-                WasmResult::Reject(s) => {
-                    panic!("Call to get_auto_finalization_status failed: {:#?}", s)
-                }
-            };
             Decode!(&result, GetAutoFinalizationStatusResponse).unwrap()
         }